requirements.txt
# Core Web Server & Gateway
fastapi
uvicorn
python-multipart
python-dotenv
httpx

# Testing & Security Auditing
pytest
pytest-asyncio
